TypeScriptize all detailed views
