+++ app/assets/javascripts/athletes.js
