+++ app/assets/javascripts/athletes/helpers/htmlHelpers.ts
@@ -13,7 +13,7 @@ namespace HtmlHelpers {
         return html;
     }
 
-    export function getLoadingIconHtml() {
+    export function getLoadingIcon() {
         let html = `
         <div class='loading-icon-panel text-center'>
             <button type='button' class='btn btn-default btn-lrg' title='Loading Data...'>
@@ -53,5 +53,118 @@ namespace HtmlHelpers {
             </div>
         </div>`;
         return html;
+    }
+
+    export function getDatatableHeaderForBestEfforts() {
+        let header = `
+            <thead>
+                <tr>
+                    <th class="col-md-1">Date</th>
+                    <th class="col-md-1 text-center badge-cell hidden-xs-down">Type</th>
+                    <th class="col-md-4">Activity</th>
+                    <th class="col-md-1">Time</th>
+                    <th class="col-md-1 hidden-xs-down">Pace</th>
+                    <th class="col-md-2 hidden-lg-down">Gear</th>
+                    <th class="col-md-1 text-center badge-cell hidden-md-down">Avg. HR</th>
+                    <th class="col-md-1 text-center badge-cell hidden-md-down">Max HR</th>
+                </tr>
+            </thead>
+        `;
+        return header;
+    }
+
+    export function getDatatableHeaderForRaces() {
+        var header = `
+            <thead>
+                <tr>
+                <th class="col-md-1">Date</th>
+                <th class="col-md-3">Activity</th>
+                <th class="col-md-1">Time</th>
+                <th class="col-md-1 hidden-xs-down">Pace</th>
+                <th class="col-md-2 hidden-lg-down">Gear</th>
+                <th class="col-md-1 hidden-md-down">Elevation</th>
+                <th class="col-md-1 hidden-md-down">Cadence</th>
+                <th class="col-md-1 text-center badge-cell hidden-md-down">Avg. HR</th>
+                <th class="col-md-1 text-center badge-cell hidden-md-down">Max HR</th>
+                </tr>
+            </thead>
+        `;
+        return header;
+    }
+
+    export function getDatatableRowForBestEfforts(item: any) {
+        let row = `
+            <tr>
+                <td>${item['start_date']}</td>
+                <td class="text-center badge-cell hidden-xs-down">
+                    <span class="label workout-type-${item['workout_type_name'].replace(/ /g, "-")}">${item['workout_type_name']}</span>
+                </td>
+                <td>
+                    <a class="strava-activity-link" href="https://www.strava.com/activities/${item['activity_id']}" target="_blank">
+                        ${item['activity_name']}
+                    </a>
+                </td>
+                <td>
+                    ${item['elapsed_time_formatted']}
+                </td>
+                <td class="hidden-xs-down">
+                    ${item['pace']}<small>${item['pace_unit']}</small>
+                </td>
+                <td class="hidden-lg-down">
+                    ${item['gear_name']}
+                </td>
+                <td class="text-center badge-cell hidden-md-down">
+                    <span class="badge ${item['average_hr_zone_class']}">
+                        ${item['average_heartrate']}
+                    </span>
+                </td>
+                <td class="text-center badge-cell hidden-md-down">
+                    <span class="badge ${item['max_hr_zone_class']}">
+                        ${item['max_heartrate']}
+                    </span>
+                </td>
+            </tr>
+        `;
+        return row;
+    }
+
+    export function getDatatableRowForRaces(item: any[]) {
+        var row = `
+            <tr>
+                <td>${item['start_date']}</td>
+                <td>
+                    <a class="strava-logo-link" href="https://www.strava.com/activities/${item['activity_id']}" target="_blank">
+                        <span></span>
+                    </a>
+                    <a href="https://www.strava.com/activities/${item['activity_id']}" target="_blank">${item['activity_name']}</a>
+                </td>
+                <td>
+                    ${item['elapsed_time_formatted']}
+                </td>
+                <td class="hidden-xs-down">
+                    ${item['pace']}<small>${item['pace_unit']}</small>
+                </td>
+                <td class="hidden-lg-down">
+                    ${item['gear_name']}
+                </td>
+                <td class="hidden-md-down">
+                    ${item['elevation']}<small> ${item['elevation_unit']}</small>
+                </td>
+                <td class="hidden-md-down">
+                    ${item['cadence']}
+                </td>
+                <td class="text-center badge-cell hidden-md-down">
+                    <span class="badge ${item['average_hr_zone_class']}">
+                        ${item['average_heartrate']}
+                    </span>
+                </td>
+                <td class="text-center badge-cell hidden-md-down">
+                    <span class="badge ${item["max_hr_zone_class"]}">
+                        ${item["max_heartrate"]}
+                    </span>
+                </td>
+            </tr>
+        `;
+        return row;
     }
 }
