--- app/assets/javascripts/athletes/races.ts
@@ -1,288 +1,0 @@
-function loadRacesByDistanceView(distanceText) {
-    var constructDataTableHtml = function(races?) {
-        var table = '<div class="row">';
-        table += '<div class="col-xs-12">';
-        table += '<div class="box">';
-        table += '<div class="box-header with-border>';
-        table += '<i class="fa fa-bar-chart-o"></i><h3 class="box-title">Data Table</h3>';
-        table += '<div class="box-body">';
-
-        if (races === undefined) {
-            table += HtmlHelpers.getLoadingIconHtml();
-        } else {
-            table += '<table class="dataTable table table-bordered table-striped">';
-            table += createRaceDatatableHeader();
-            table += '<tbody>';
-
-            races.forEach(function(item) {
-                table += createRaceDatatableRow(item);
-            });
-
-            table += '</tbody>';
-            table += '</table>';
-        }
-        table += '</div></div></div></div></div>';
-        return table;
-    };
-    var prepareView = function () {
-        AppHelpers.setContentHeader("Races - " + distanceText);
-        AppHelpers.appendToPageTitle(' |  Races  - ' + distanceText);
-
-        AppHelpers.resetNavigationItems();
-        var navigationAnchor = $("a[id^='races-for-distance-" + distanceText.toLowerCase().replace(/ /g, '-').replace(/\//g, '-') + "']");
-        AppHelpers.setNavigationItem(navigationAnchor);
-
-        var mainContent = $('#main-content');
-        mainContent.empty(); // Empty main content.
-
-        // Create empty progression chart with loading icon.
-        var showLoadingIcon = true;
-        var topRow = '<div class="row">';
-        topRow += constructChartHtml('progression-chart', 'Progression Chart', 8, showLoadingIcon);
-        topRow += constructChartHtml('year-distribution-pie-chart', 'Year Distribution Chart', 4, showLoadingIcon);
-        topRow += '</div>';
-        mainContent.append(topRow);
-
-        // Create empty data table with loading icon.
-        var table = constructDataTableHtml();
-        mainContent.append(table);
-
-        // Create empty pie charts with loading icon.
-        var pieCharts = '<div class="row">';
-        pieCharts += constructChartHtml('gear-count-chart', 'Gear Count Chart', 6, showLoadingIcon);
-        pieCharts += constructChartHtml('gear-mileage-chart', 'Gear Mileage Chart', 6, showLoadingIcon);
-        pieCharts += '</div>';
-        mainContent.append(pieCharts);
-    };
-    var createView = function() {
-        var distance = distanceText.trim().replace(/\//g, '|').replace(/\s/g, '-').toLowerCase();
-        AppHelpers.pushStateToWindow(AppHelpers.getBaseUrl() + '/races/' + distance);
-        $.ajax({
-            url: AppHelpers.getApiBaseUrl() + '/races/' + distance,
-            dataType: 'json',
-            async: false,
-            success: function(data) {
-
-                var races = [];
-                $.each(data, function(key, value) {
-                    races.push(value);
-                });
-
-                var mainContent = $('#main-content');
-                mainContent.empty();
-
-                // Create a progression chart when distance is not 'Other'.
-                var progressionChartId = 'progression-chart';
-                var topRow = '<div class="row">';
-                topRow += constructChartHtml(progressionChartId, 'Progression Chart', 8, false);
-                topRow += constructChartHtml('year-distribution-pie-chart', 'Year Distribution Chart', 4, false);
-                topRow += '</div>';
-                mainContent.append(topRow);
-                if (distanceText === 'Other') {
-                    createChartMessage(progressionChartId, "Not Applicable");
-                } else {
-                    createProgressionChart(progressionChartId, races);
-                }
-                createYearDistributionChart('year-distribution-pie-chart', races);
-
-                // Create data table.
-                var table = constructDataTableHtml(races);
-                mainContent.append(table);
-                var setupDataTable = function() {
-                    $(".dataTable").each(function() {
-                        $(this).DataTable({
-                            "columnDefs": [{
-                                "targets": [2, 3, 5, 6, 7, 8], // Disable searching for Time, Pace, Elevation, Cadence and HRs.
-                                "searchable": false
-                            }],
-                            "iDisplayLength": 10,
-                            "order": [
-                                [0, "desc"]
-                            ]
-                        });
-                    });
-                };
-                setupDataTable();
-
-                var pieCharts = '<div class="row">';
-                pieCharts += constructChartHtml('gear-count-chart', 'Gear Count Chart', 6, false);
-                pieCharts += constructChartHtml('gear-mileage-chart', 'Gear Mileage Chart', 6, false);
-                pieCharts += '</div>';
-                mainContent.append(pieCharts);
-                createGearCountChart('gear-count-chart', races);
-                createGearMileageChart('gear-mileage-chart', races);
-            }
-        });
-    };
-
-    prepareView();
-    createView();
-}
-
-function loadRacesByYearView(year) {
-    var constructDataTableHtml = function(races?) {
-        var table = '<div class="row">';
-        table += '<div class="col-xs-12">';
-        table += '<div class="box">';
-        table += '<div class="box-header with-border>';
-        table += '<i class="fa fa-bar-chart-o"></i><h3 class="box-title">Data Table</h3>';
-        table += '<div class="box-body">';
-
-        if (races === undefined) {
-            table += HtmlHelpers.getLoadingIconHtml();
-        } else {
-            var allDistances = [
-                '100 miles', '100k', '50 miles', '50k', 'Marathon', 'Half Marathon',
-                '20k', '15k', '10k', '5k', '3000m', '1 mile', 'Other'
-            ]; // Just hard code race distances here. No need to get from server side for now.
-            var distancesToDisplay = [];
-            allDistances.forEach(function(distance) {
-                for (var index = 0; index < races.length; ++index) {
-                    var raceDistance = races[index]["race_distance"];
-                    if (distance === raceDistance && distancesToDisplay.indexOf(raceDistance) === -1) {
-                        distancesToDisplay.push(raceDistance);
-                        break;
-                    }
-                }
-            });
-            distancesToDisplay.forEach(function(distance) {
-                table += '<h4>' + distance + '</h4>';
-                table += '<table class="dataTable table table-bordered table-striped">';
-                table += createRaceDatatableHeader();
-                table += '<tbody>';
-                races.forEach(function(item) {
-                    if (distance === item['race_distance']) {
-                        table += createRaceDatatableRow(item);
-                    }
-                });
-                table += '</tbody>';
-                table += '</table>';
-            });
-        }
-        table += '</div></div></div></div></div>';
-        return table;
-    };
-    var prepareView = function () {
-        AppHelpers.setContentHeader("Races - " + year);
-        AppHelpers.appendToPageTitle(' |  Races  - ' + year);
-
-        AppHelpers.resetNavigationItems();
-        var navigationAnchor = $("a[id^='races-for-year-" + year + "']");
-        AppHelpers.setNavigationItem(navigationAnchor);
-
-        var mainContent = $('#main-content');
-        mainContent.empty(); // Empty main content.
-
-        // Create an empty monthly distribution chart with loading icon.
-        var showLoadingIcon = true;
-        var barCharts = '<div class="row">';
-        barCharts += constructChartHtml('distances-distribution-chart', 'Distance Distribution Chart', 6, showLoadingIcon);
-        barCharts += constructChartHtml('monthly-distribution-chart', 'Monthly Distribution Chart', 6, showLoadingIcon);
-        barCharts += '</div>';
-        mainContent.append(barCharts);
-
-        // Create empty data table with loading icon.
-        var table = constructDataTableHtml();
-        mainContent.append(table);
-
-        // Create empty pie charts with loading icon.
-        var pieCharts = '<div class="row">';
-        pieCharts += constructChartHtml('gear-count-chart', 'Gear Count Chart', 6, showLoadingIcon);
-        pieCharts += constructChartHtml('gear-mileage-chart', 'Gear Mileage Chart', 6, showLoadingIcon);
-        pieCharts += '</div>';
-        mainContent.append(pieCharts);
-    };
-    var createView = function () {
-        AppHelpers.pushStateToWindow(AppHelpers.getBaseUrl() + '/races/' + year);
-        $.ajax({
-            url: AppHelpers.getApiBaseUrl() + '/races/' + year,
-            dataType: 'json',
-            async: false,
-            success: function(data) {
-
-                var races = [];
-                $.each(data, function(key, value) {
-                    races.push(value);
-                });
-
-                var mainContent = $('#main-content');
-                mainContent.empty();
-
-                var barCharts = '<div class="row">';
-                barCharts += constructChartHtml('distances-distribution-chart', 'Distance Distribution Chart', 6, false);
-                barCharts += constructChartHtml('month-distribution-chart', 'Month Distribution Chart', 6, false);
-                mainContent.append(barCharts);
-                createRaceDistancesChart('distances-distribution-chart', races);
-                createMonthDistributionChart('month-distribution-chart', races);
-
-                // Create data table.
-                var table = constructDataTableHtml(races);
-                mainContent.append(table);
-                var setupDataTable = function() {
-                    $(".dataTable").each(function() {
-                        $(this).DataTable({
-                            "bFilter": false,
-                            "bPaginate": false,
-                            "iDisplayLength": 10,
-                            "info": false,
-                            "order": [
-                                [0, "desc"]
-                            ]
-                        });
-                    });
-                };
-                setupDataTable();
-
-                var pieCharts = '<div class="row">';
-                pieCharts += constructChartHtml('gear-count-chart', 'Gear Count Chart', 6, false);
-                pieCharts += constructChartHtml('gear-mileage-chart', 'Gear Mileage Chart', 6, false);
-                pieCharts += '</div>';
-                mainContent.append(pieCharts);
-                createGearCountChart('gear-count-chart', races);
-                createGearMileageChart('gear-mileage-chart', races);
-            }
-        });
-    };
-
-    prepareView();
-    createView();
-}
-
-function createRaceDatatableHeader() {
-    var header = '<thead><tr>';
-    header += '<th class="col-md-1">Date</th>';
-    header += '<th class="col-md-3">Activity</th>';
-    header += '<th class="col-md-1">Time</th>';
-    header += '<th class="col-md-1 hidden-xs-down">Pace</th>';
-    header += '<th class="col-md-2 hidden-lg-down">Gear</th>';
-    header += '<th class="col-md-1 hidden-md-down">Elevation</th>';
-    header += '<th class="col-md-1 hidden-md-down">Cadence</th>';
-    header += '<th class="col-md-1 text-center badge-cell hidden-md-down">Avg. HR</th>';
-    header += '<th class="col-md-1 text-center badge-cell hidden-md-down">Max HR</th>';
-    header += '</tr></thead>';
-    return header;
-}
-
-function createRaceDatatableRow(item) {
-    var row = '<tr>';
-    row += '<td>' + item['start_date'] + '</td>';
-    row += '<td>';
-    row += '<a class="strava-logo-link" href="https://www.strava.com/activities/' + item["activity_id"] +
-        '" target="_blank"><span></span></a>';
-    row += '<a href="https://www.strava.com/activities/' + item["activity_id"] +
-        '" target="_blank">' + item["activity_name"] + '</a>';
-    row += '</td>';
-    row += '<td>' + item["elapsed_time_formatted"] + '</td>';
-    row += '<td class="hidden-xs-down">' + item["pace"] + '<small>' + item["pace_unit"] + '</small></td>';
-    row += '<td class="hidden-lg-down">' + item["gear_name"] + '</td>';
-    row += '<td class=" hidden-md-down">' + item["elevation"] + '<small> ' + item["elevation_unit"] + '</small></td>';
-    row += '<td class=" hidden-md-down">' + item["cadence"] + '</td>';
-    row += '<td class="text-center badge-cell hidden-md-down">';
-    row += '<span class="badge ' + item["average_hr_zone_class"] + '">' + item["average_heartrate"] + '</span>';
-    row += '</td>';
-    row += '<td class="text-center badge-cell hidden-md-down">';
-    row += '<span class="badge ' + item["max_hr_zone_class"] + '">' + item["max_heartrate"] + '</span>';
-    row += '</td>';
-    row += '</tr>';
-    return row;
-}
